added missing files from other branches
